+++ temba/orgs/views/views.py
@@ -78,9 +78,9 @@ from ..models import (
     User,
     UserSettings,
 )
-from .base import BaseMenuView
+from .base import BaseDeleteModal, BaseListView, BaseMenuView
 from .forms import SignupForm, SMTPForm
-from .mixins import InferOrgMixin, OrgObjPermsMixin, OrgPermsMixin
+from .mixins import InferOrgMixin, OrgObjPermsMixin, OrgPermsMixin, RequireFeatureMixin
 
 # session key for storing a two-factor enabled user's id once we've checked their password
 TWO_FACTOR_USER_SESSION_KEY = "_two_factor_user_id"
@@ -338,6 +338,9 @@ class InferUserMixin:
 class UserCRUDL(SmartCRUDL):
     model = User
     actions = (
+        "list",
+        "update",
+        "delete",
         "edit",
         "forget",
         "recover",
@@ -349,6 +352,100 @@ class UserCRUDL(SmartCRUDL):
         "verify_email",
         "send_verification_email",
     )
+
+    class List(SpaMixin, RequireFeatureMixin, ContextMenuMixin, OrgPermsMixin, SmartListView):
+        require_feature = Org.FEATURE_USERS
+        title = _("Users")
+        menu_path = "/settings/users"
+        search_fields = ("email__icontains", "first_name__icontains", "last_name__icontains")
+
+        def derive_queryset(self, **kwargs):
+            return (
+                super()
+                .derive_queryset(**kwargs)
+                .filter(id__in=self.request.org.get_users().values_list("id", flat=True))
+                .order_by(Lower("email"))
+            )
+
+        def get_context_data(self, **kwargs):
+            context = super().get_context_data(**kwargs)
+
+            for user in context["object_list"]:
+                user.role = self.request.org.get_user_role(user)
+
+            context["has_viewers"] = self.request.org.get_users(roles=[OrgRole.VIEWER]).exists()
+            context["admin_count"] = self.request.org.get_users(roles=[OrgRole.ADMINISTRATOR]).count()
+            return context
+
+    class Update(RequireFeatureMixin, ModalFormMixin, OrgObjPermsMixin, SmartUpdateView):
+        class Form(forms.ModelForm):
+            role = forms.ChoiceField(choices=OrgRole.choices(), required=True, label=_("Role"), widget=SelectWidget())
+
+            class Meta:
+                model = User
+                fields = ("role",)
+
+        form_class = Form
+        require_feature = Org.FEATURE_USERS
+
+        def get_object_org(self):
+            return self.request.org
+
+        def get_queryset(self):
+            return self.request.org.get_users()
+
+        def derive_initial(self):
+            # viewers default to editors
+            role = self.request.org.get_user_role(self.object)
+            return {"role": OrgRole.EDITOR.code if role == OrgRole.VIEWER else role.code}
+
+        def save(self, obj):
+            role = OrgRole.from_code(self.form.cleaned_data["role"])
+
+            # don't update if user is the last administrator and role is being changed to something else
+            has_other_admins = self.request.org.get_users(roles=[OrgRole.ADMINISTRATOR]).exclude(id=obj.id).exists()
+            if role != OrgRole.ADMINISTRATOR and not has_other_admins:
+                return obj
+
+            self.request.org.add_user(obj, role)
+            return obj
+
+        def get_success_url(self):
+            return reverse("orgs.user_list") if self.has_org_perm("orgs.user_list") else reverse("orgs.org_start")
+
+    class Delete(RequireFeatureMixin, OrgObjPermsMixin, SmartDeleteView):
+        permission = "orgs.user_update"
+        require_feature = Org.FEATURE_USERS
+        fields = ("id",)
+        submit_button_name = _("Remove")
+        cancel_url = "@orgs.user_list"
+        redirect_url = "@orgs.user_list"
+
+        def get_object_org(self):
+            return self.request.org
+
+        def get_queryset(self):
+            return self.request.org.get_users()
+
+        def get_context_data(self, **kwargs):
+            context = super().get_context_data(**kwargs)
+            context["submit_button_name"] = self.submit_button_name
+            return context
+
+        def post(self, request, *args, **kwargs):
+            user = self.get_object()
+
+            # only actually remove user if they're not the last administator
+            if self.request.org.get_users(roles=[OrgRole.ADMINISTRATOR]).exclude(id=user.id).exists():
+                self.request.org.remove_user(user)
+
+            return HttpResponseRedirect(self.get_redirect_url())
+
+        def get_redirect_url(self):
+            still_in_org = self.request.org.has_user(self.request.user) or self.request.user.is_staff
+
+            # if current user no longer belongs to this org, redirect to org chooser
+            return reverse("orgs.user_list") if still_in_org else reverse("orgs.org_choose")
 
     class Forget(SmartFormView):
         class Form(forms.Form):
@@ -857,8 +954,6 @@ class OrgCRUDL(SmartCRUDL):
         "grant",
         "choose",
         "delete_child",
-        "manage_accounts",
-        "manage_accounts_sub_org",
         "menu",
         "country",
         "languages",
@@ -924,21 +1019,27 @@ class OrgCRUDL(SmartCRUDL):
                         )
                     )
 
-                if self.has_org_perm("orgs.org_manage_accounts") and Org.FEATURE_USERS in org.features:
-                    menu.append(
-                        self.create_menu_item(
-                            name=_("Users"),
-                            icon="users",
-                            href="orgs.org_manage_accounts",
-                            count=org.users.count(),
-                        )
-                    )
-
                 menu.append(self.create_menu_item(name=_("Resthooks"), icon="resthooks", href="orgs.org_resthooks"))
 
                 if self.has_org_perm("notifications.incident_list"):
                     menu.append(
                         self.create_menu_item(name=_("Incidents"), icon="incidents", href="notifications.incident_list")
+                    )
+
+                if Org.FEATURE_USERS in org.features and self.has_org_perm("orgs.user_list"):
+                    menu.append(self.create_divider())
+                    menu.append(
+                        self.create_menu_item(
+                            name=_("Users"), icon="users", href="orgs.user_list", count=org.users.count()
+                        )
+                    )
+                    menu.append(
+                        self.create_menu_item(
+                            name=_("Invitations"),
+                            icon="invitations",
+                            href="orgs.invitation_list",
+                            count=org.invitations.filter(is_active=True).count(),
+                        )
                     )
 
                 menu.append(self.create_divider())
@@ -1349,184 +1450,6 @@ class OrgCRUDL(SmartCRUDL):
             self.object.release(request.user)
             return self.render_modal_response()
 
-    class ManageAccounts(SpaMixin, InferOrgMixin, ContextMenuMixin, OrgPermsMixin, SmartUpdateView):
-        class AccountsForm(forms.ModelForm):
-            def __init__(self, org, *args, **kwargs):
-                super().__init__(*args, **kwargs)
-
-                self.org = org
-                self.user_rows = []
-                self.invite_rows = []
-                self.add_per_user_fields(org)
-                self.add_per_invite_fields(org)
-
-            def add_per_user_fields(self, org: Org):
-                role_choices = [(r.code, r.display) for r in (OrgRole.ADMINISTRATOR, OrgRole.EDITOR, OrgRole.AGENT)]
-                role_choices_inc_viewer = role_choices + [(OrgRole.VIEWER.code, OrgRole.VIEWER.display)]
-
-                for user in org.users.order_by("email"):
-                    role = org.get_user_role(user)
-
-                    role_field = forms.ChoiceField(
-                        choices=role_choices_inc_viewer if role == OrgRole.VIEWER else role_choices,
-                        required=True,
-                        initial=role.code,
-                        label=" ",
-                        widget=SelectWidget(),
-                    )
-                    remove_field = forms.BooleanField(
-                        required=False, label=" ", widget=CheckboxWidget(attrs={"widget_only": True})
-                    )
-
-                    self.fields.update(
-                        OrderedDict([(f"user_{user.id}_role", role_field), (f"user_{user.id}_remove", remove_field)])
-                    )
-                    self.user_rows.append(
-                        {"user": user, "role_field": f"user_{user.id}_role", "remove_field": f"user_{user.id}_remove"}
-                    )
-
-            def add_per_invite_fields(self, org: Org):
-                for invite in org.invitations.filter(is_active=True).order_by("email"):
-                    role_field = forms.ChoiceField(
-                        choices=[(r.code, r.display) for r in (OrgRole.ADMINISTRATOR, OrgRole.EDITOR, OrgRole.AGENT)],
-                        required=True,
-                        initial=invite.role.code,
-                        label=" ",
-                        widget=SelectWidget(),
-                        disabled=True,
-                    )
-                    remove_field = forms.BooleanField(
-                        required=False, label=" ", widget=CheckboxWidget(attrs={"widget_only": True})
-                    )
-
-                    self.fields.update(
-                        OrderedDict(
-                            [(f"invite_{invite.id}_role", role_field), (f"invite_{invite.id}_remove", remove_field)]
-                        )
-                    )
-                    self.invite_rows.append(
-                        {
-                            "invite": invite,
-                            "role_field": f"invite_{invite.id}_role",
-                            "remove_field": f"invite_{invite.id}_remove",
-                        }
-                    )
-
-            def get_submitted_roles(self) -> dict:
-                """
-                Returns a dict of users to roles from the current form data. None role means removal.
-                """
-                roles = {}
-
-                for row in self.user_rows:
-                    role = self.cleaned_data.get(row["role_field"])
-                    remove = self.cleaned_data.get(row["remove_field"])
-                    roles[row["user"]] = OrgRole.from_code(role) if not remove else None
-                return roles
-
-            def get_submitted_invite_removals(self) -> list:
-                """
-                Returns a list of invites to be removed.
-                """
-                invites = []
-                for row in self.invite_rows:
-                    if self.cleaned_data[row["remove_field"]]:
-                        invites.append(row["invite"])
-                return invites
-
-            def clean(self):
-                super().clean()
-
-                new_roles = self.get_submitted_roles()
-                has_admin = False
-                for new_role in new_roles.values():
-                    if new_role == OrgRole.ADMINISTRATOR:
-                        has_admin = True
-                        break
-
-                if not has_admin:
-                    raise forms.ValidationError(_("A workspace must have at least one administrator."))
-
-            class Meta:
-                model = Invitation
-                fields = ()
-
-        form_class = AccountsForm
-        success_url = "@orgs.org_manage_accounts"
-        title = _("Users")
-        menu_path = "/settings/users"
-
-        def pre_process(self, request, *args, **kwargs):
-            if Org.FEATURE_USERS not in request.org.features:
-                return HttpResponseRedirect(reverse("orgs.org_workspace"))
-
-        def build_context_menu(self, menu):
-            menu.add_modax(_("Invite"), "invite-create", reverse("orgs.invitation_create"), as_button=True)
-
-        def get_form_kwargs(self):
-            kwargs = super().get_form_kwargs()
-            kwargs["org"] = self.get_object()
-            return kwargs
-
-        def post_save(self, obj):
-            obj = super().post_save(obj)
-            org = self.get_object()
-
-            # delete any invitations which have been checked for removal
-            for invite in self.form.get_submitted_invite_removals():
-                org.invitations.filter(id=invite.id).delete()
-
-            # update org users with new roles
-            for user, new_role in self.form.get_submitted_roles().items():
-                if not new_role:
-                    org.remove_user(user)
-                elif org.get_user_role(user) != new_role:
-                    org.add_user(user, new_role)
-
-            return obj
-
-        def get_context_data(self, **kwargs):
-            context = super().get_context_data(**kwargs)
-            org = self.get_object()
-            context["org"] = org
-            context["has_invites"] = org.invitations.filter(is_active=True).exists()
-            context["has_viewers"] = org.get_users(roles=[OrgRole.VIEWER]).exists()
-            return context
-
-        def get_success_url(self):
-            still_in_org = self.get_object().has_user(self.request.user) or self.request.user.is_staff
-
-            # if current user no longer belongs to this org, redirect to org chooser
-            return reverse("orgs.org_manage_accounts") if still_in_org else reverse("orgs.org_choose")
-
-    class ManageAccountsSubOrg(ManageAccounts):
-        menu_path = "/settings/workspaces"
-
-        def pre_process(self, request, *args, **kwargs):
-            pass
-
-        def build_context_menu(self, menu):
-            menu.add_modax(
-                _("Invite"),
-                "invite-create",
-                reverse("orgs.invitation_create") + f"?org={self.target_org.id}",
-                as_button=True,
-            )
-
-        def derive_title(self):
-            return self.target_org.name
-
-        def get_object(self, *args, **kwargs):
-            return self.target_org
-
-        @cached_property
-        def target_org(self):
-            return get_object_or_404(self.request.org.children.filter(id=int(self.request.GET.get("org", 0))))
-
-        def get_success_url(self):  # pragma: needs cover
-            org_id = self.request.GET.get("org")
-            return "%s?org=%s" % (reverse("orgs.org_manage_accounts_sub_org"), org_id)
-
     class Service(StaffOnlyMixin, SmartFormView):
         class ServiceForm(forms.Form):
             other_org = ModelChoiceField(queryset=Org.objects.all(), widget=forms.HiddenInput())
@@ -1578,15 +1501,7 @@ class OrgCRUDL(SmartCRUDL):
 
             return queryset.filter(id__in=ids, is_active=True).order_by("-parent", "name")
 
-        def get_context_data(self, **kwargs):
-            context = super().get_context_data(**kwargs)
-            org = self.get_object()
-            if self.has_org_perm("orgs.org_manage_accounts") and Org.FEATURE_USERS in org.features:
-                context["manage_users"] = True
-
-            return context
-
-    class Create(NonAtomicMixin, ModalFormMixin, InferOrgMixin, OrgPermsMixin, SmartCreateView):
+    class Create(NonAtomicMixin, RequireFeatureMixin, ModalFormMixin, InferOrgMixin, OrgPermsMixin, SmartCreateView):
         class Form(forms.ModelForm):
             TYPE_CHILD = "child"
             TYPE_NEW = "new"
@@ -1607,12 +1522,7 @@ class OrgCRUDL(SmartCRUDL):
                 fields = ("type", "name", "timezone")
 
         form_class = Form
-
-        def pre_process(self, request, *args, **kwargs):
-            # if org has neither feature then redirect
-            features = self.request.org.features
-            if Org.FEATURE_NEW_ORGS not in features and Org.FEATURE_CHILD_ORGS not in features:
-                return HttpResponseRedirect(reverse("orgs.org_workspace"))
+        require_feature = (Org.FEATURE_NEW_ORGS, Org.FEATURE_CHILD_ORGS)
 
         def get_form_kwargs(self):
             kwargs = super().get_form_kwargs()
@@ -2211,15 +2121,30 @@ class OrgCRUDL(SmartCRUDL):
 
 class InvitationCRUDL(SmartCRUDL):
     model = Invitation
-    actions = ("create",)
+    actions = ("list", "create", "delete")
 
-    class Create(ModalFormMixin, OrgPermsMixin, SmartCreateView):
+    class List(RequireFeatureMixin, ContextMenuMixin, BaseListView):
+        require_feature = Org.FEATURE_USERS
+        title = _("Invitations")
+        menu_path = "/settings/invitations"
+        default_order = ("-created_on",)
+
+        def build_context_menu(self, menu):
+            menu.add_modax(_("New"), "invite-create", reverse("orgs.invitation_create"), as_button=True)
+
+        def derive_queryset(self, **kwargs):
+            return super().derive_queryset(**kwargs).filter(is_active=True)
+
+        def get_context_data(self, **kwargs):
+            context = super().get_context_data(**kwargs)
+            context["validity_days"] = settings.INVITATION_VALIDITY.days
+            return context
+
+    class Create(RequireFeatureMixin, ModalFormMixin, OrgPermsMixin, SmartCreateView):
         class Form(forms.ModelForm):
-            ROLE_CHOICES = [(r.code, r.display) for r in (OrgRole.AGENT, OrgRole.EDITOR, OrgRole.ADMINISTRATOR)]
-
             email = forms.EmailField(widget=InputWidget(attrs={"widget_only": True, "placeholder": _("Email Address")}))
             role = forms.ChoiceField(
-                choices=ROLE_CHOICES, initial=OrgRole.EDITOR.code, label=_("Role"), widget=SelectWidget()
+                choices=OrgRole.choices(), initial=OrgRole.EDITOR.code, label=_("Role"), widget=SelectWidget()
             )
 
             def __init__(self, org, *args, **kwargs):
@@ -2243,20 +2168,14 @@ class InvitationCRUDL(SmartCRUDL):
                 fields = ("email", "role")
 
         form_class = Form
+        require_feature = Org.FEATURE_USERS
         title = ""
         submit_button_name = _("Send")
-        success_url = "@orgs.org_manage_accounts"
-
-        def get_dest_org(self):
-            org_id = self.request.GET.get("org")
-            if org_id:
-                return get_object_or_404(self.request.org.children.filter(id=org_id))
-
-            return self.request.org
+        success_url = "@orgs.invitation_list"
 
         def get_form_kwargs(self):
             kwargs = super().get_form_kwargs()
-            kwargs["org"] = self.get_dest_org()
+            kwargs["org"] = self.request.org
             return kwargs
 
         def get_context_data(self, **kwargs):
@@ -2264,20 +2183,20 @@ class InvitationCRUDL(SmartCRUDL):
             context["validity_days"] = settings.INVITATION_VALIDITY.days
             return context
 
-        def pre_save(self, obj):
-            org = self.get_dest_org()
-
-            assert Org.FEATURE_USERS in org.features
-
-            obj.org = org
-            obj.user_group = self.form.cleaned_data["role"]
-
-            return super().pre_save(obj)
+        def save(self, obj):
+            self.object = Invitation.create(
+                self.request.org, self.request.user, obj.email, OrgRole.from_code(self.form.cleaned_data["role"])
+            )
 
         def post_save(self, obj):
             obj.send()
 
             return super().post_save(obj)
+
+    class Delete(RequireFeatureMixin, BaseDeleteModal):
+        require_feature = Org.FEATURE_USERS
+        cancel_url = "@orgs.invitation_list"
+        redirect_url = "@orgs.invitation_list"
 
 
 class OrgImportCRUDL(SmartCRUDL):