+++ temba/settings_common.py
@@ -385,8 +385,6 @@ PERMISSIONS = {
         "join_accept",
         "join",
         "languages",
-        "manage_accounts_sub_org",
-        "manage_accounts",
         "manage_integrations",
         "manage",
         "menu",
@@ -496,8 +494,6 @@ GROUP_PERMISSIONS = {
         "orgs.org_export",
         "orgs.org_flow_smtp",
         "orgs.org_languages",
-        "orgs.org_manage_accounts_sub_org",
-        "orgs.org_manage_accounts",
         "orgs.org_manage_integrations",
         "orgs.org_menu",
         "orgs.org_prometheus",
@@ -508,6 +504,7 @@ GROUP_PERMISSIONS = {
         "orgs.orgimport.*",
         "orgs.user_list",
         "orgs.user_tokens",
+        "orgs.user_update",
         "request_logs.httplog_list",
         "request_logs.httplog_read",
         "request_logs.httplog_webhooks",
@@ -589,7 +586,6 @@ GROUP_PERMISSIONS = {
         "orgs.org_resthooks",
         "orgs.org_workspace",
         "orgs.orgimport.*",
-        "orgs.user_list",
         "orgs.user_tokens",
         "request_logs.httplog_webhooks",
         "templates.template_list",
@@ -656,7 +652,6 @@ GROUP_PERMISSIONS = {
         "orgs.org_menu",
         "orgs.org_read",
         "orgs.org_workspace",
-        "orgs.user_list",
         "templates.template_list",
         "templates.template_read",
         "tickets.ticket_export",
@@ -684,6 +679,8 @@ GROUP_PERMISSIONS = {
 
 # extra permissions that only apply to API requests (wildcard notation not supported here)
 API_PERMISSIONS = {
+    "Editors": ("orgs.user_list",),
+    "Viewers": ("orgs.user_list",),
     "Agents": (
         "contacts.contact_create",
         "contacts.contact_list",
@@ -695,7 +692,7 @@ API_PERMISSIONS = {
         "msgs.msg_create",
         "orgs.org_read",
         "orgs.user_list",
-    )
+    ),
 }
 
 # -----------------------------------------------------------------------------------